Refactor oneshot mode state into a separate module
